Mods to install logic
Allow yamlcpp to be installed even if it is included as part of another cmake project.

Also, allow to set a prefix to append to CMAKE_INSTALL_PREFIX for yamlcpp.
This will install, say, yamlcpp headers in
  /<prefix>/include
which allows to avoid mixing yamlcpp include/libs/cmake files with the ones
of the host project.